siso: preserve directory mtime/cmdhash if it is used as output
ninja checks mtime and cmdhash of output whether if it needs to
run the step or not in incremental build, so we preserve this
in .siso_fs_state.
Some steps use directory as output, so we need to preserve
mtime and cmdhash for such directories for incremental build.

Bug: 1451725, b/254182269
Change-Id: I8ce858ceaedaa0d8ef8eb60b2d8bc4a068677d43
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4584493
Reviewed-by: Junji Watanabe <jwata@google.com>
Auto-Submit: Fumitoshi Ukai <ukai@google.com>
Commit-Queue: Fumitoshi Ukai <ukai@google.com>
Cr-Commit-Position: refs/heads/main@{#57254}
GitOrigin-RevId: 1cfb1835e358adf248dad24eacd60042e21865e7