siso: fix missing inputs for remote link
remote link sometimes failed due to missing inputs.
This is because that alink step that solink step depends
was skipped as it was up-to-date by the previous build,
so edgeRule for the alink was not set and accumulate didn't
work to add *.o from *.a thin archive.

Since stepConfig.Lookup is not fast enough to check
it always for no-op build, lazily lookup only when
it is required.

Bug: b/322449841
Change-Id: I61f4541522958b3acbf6e8059c20a647f2aa9c32
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5237655
Commit-Queue: Junji Watanabe <jwata@google.com>
Auto-Submit: Fumitoshi Ukai <ukai@google.com>
Reviewed-by: Junji Watanabe <jwata@google.com>
Cr-Commit-Position: refs/heads/main@{#62747}
GitOrigin-RevId: de759a1d10af25bac7e57f9073d1fa30c05e998d