siso: make foo.h^ equivalent with foo.cc^
When header file "foo.h^" is given as target,
but "foo.h" is not in build graph, it failed.

It would be convenient to make it equivalent with "foo.cc^",
i.e. if "foo.h" is not in build graph, use its source ("foo.cc")
instead to lookup the targets.

Bug: b/335792430
Change-Id: I529010eb27533de498b4e597117f7acc59757d18
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5469929
Auto-Submit: Fumitoshi Ukai <ukai@google.com>
Reviewed-by: Junji Watanabe <jwata@google.com>
Commit-Queue: Fumitoshi Ukai <ukai@google.com>
Cr-Commit-Position: refs/heads/main@{#64684}
GitOrigin-RevId: 895dfa2dc7f93792df8a332826a0af2e6b015b3c