siso: fix crash in lazyCompute: "send no closed channel".
When interrupted, build finished and hashfs is closed,
but some goroutines for scandeps keep running, and
it would trigger lazyCompute and try to send on closed channel.

Set digester's q to nil when closed,
and check it in lazyCompute before sending to the channel.

Also check whether ctx is done in lazyCompute and compute.

Bug: b/347635033
Change-Id: I5ef43467264b8621aebb12936773f93a9d6f65ef
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5636173
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Auto-Submit: Fumitoshi Ukai <ukai@google.com>
Commit-Queue: Fumitoshi Ukai <ukai@google.com>
Cr-Commit-Position: refs/heads/main@{#66253}
GitOrigin-RevId: 44ccac1954faabf1f22fac0df8c214d3eaa76f3d