siso: webui outdir in urls
this prepares the webui for supporting outdir switching, but doesn't
implement outdir switching yet.

explicit 404 in the root handler is removed, but this is fine because
the outdir http handlers don't have a catch-all handler. hence any
urls that don't match a handler will continue to 404.

Bug: 349287453
Change-Id: I01662d068aea048ff142912cb785000862587c14
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5701740
Reviewed-by: Junji Watanabe <jwata@google.com>
Commit-Queue: Richard Wang <richardwa@google.com>
Cr-Commit-Position: refs/heads/main@{#66767}
GitOrigin-RevId: fc8415bd845ec4bf705027fb533f91af0095b19b