siso: don't enable zstd by default yet
Older versions of Siso are not able to read zstd compressed state files
and will silently ignore changes to the input files from then on. Users
might accidentally run into this when working on multiple branches, of
which some are still synced to a commit that uses an earlier version of
Siso. We will need to find a solution for this before enabling zstd by
default.

Change-Id: Ic068d5b68508558dd7431946787cde6f6539d633
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5759123
Reviewed-by: Fumitoshi Ukai <ukai@google.com>
Commit-Queue: Philipp Wollermann <philwo@google.com>
Cr-Commit-Position: refs/heads/main@{#67354}
GitOrigin-RevId: d53d7d75098abdd48b89b1fb4fb4acc44a64a1dd