siso: use BGZF to (de)compress fs_state
BGZF (Blocked GNU Zip Format) is a specialized variation of the GZIP
compression format designed to improve efficiency when working with
large files. Its origins lie in the field of bioinformatics, where it is
used to compress DNA sequencing data in the SAM/BAM file formats.

The key feature of BGZF is that it divides the input data into blocks of
up to 64 KB in size, and compresses each block independently using GZIP.
This creates a series of concatenated GZIP blocks within the compressed
file. Because the original GZIP file format allows for the concatenation
of multiple GZIP streams, BGZF files are fully compatible with existing
GZIP tools and libraries.

This block-wise approach enables parallel processing of blocks during
both compression and decompression, speeding up both operations
significantly on multi-core processors.

The specification for BGZF is available at the following URL:
https://samtools.github.io/hts-specs/SAMv1.pdf, section 4.1.

Change-Id: Ica87752b319bed2fba508df17db5ae053f56cd39
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5758370
Commit-Queue: Philipp Wollermann <philwo@google.com>
Reviewed-by: Junji Watanabe <jwata@google.com>
Cr-Commit-Position: refs/heads/main@{#67387}
GitOrigin-RevId: 2485fb468e55d27a3a2c0d78e54b35cb16acf52f