siso: webui first pass making a proper build step page
The old prototype page just output all of the raw stats as a list.
This is a first pass to present those stats in a more readable way.
Raw stats are still accessible for now but collapsed.

We also scan for other builds the step was found to have run in, so
we can show and link to the step in these other builds. (This relies
on the output name, so it's not 100% guaranteed to stay the same if
this output changes across builds. Hence the "found in" wording.)

Bug: 361461051
Change-Id: Id825f13e11a8094c57fd7cdfb413099d0c0750ff
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5850239
Reviewed-by: Philipp Wollermann <philwo@google.com>
Commit-Queue: Richard Wang <richardwa@google.com>
Cr-Commit-Position: refs/heads/main@{#68300}
GitOrigin-RevId: 58c71fad9b1e097dd7cb07e68800f54a81244517