siso: cleandead ignores additional outputs
Siso recorded command outputs in hashfs,
and cleandead would remove files if previously generated file
is not appeared in build graph.

If siso config adds outputs, it would not be appeared in build graph,
but recorded as outputs of command,
So cleandead would remove the files,
and it caused the issue when we enabled remote proguard actions.

To mitigate this, don't record additional outputs as command output
if it is not listed in build graph, so cleandead will ignore
such additional outputs.

Bug: b/371928116
Change-Id: Iff4e9c7ae8fa67519d9bdb2af28ba20059fccab2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5914416
Reviewed-by: Junji Watanabe <jwata@google.com>
Commit-Queue: Junji Watanabe <jwata@google.com>
Auto-Submit: Fumitoshi Ukai <ukai@google.com>
Commit-Queue: Fumitoshi Ukai <ukai@google.com>
Cr-Commit-Position: refs/heads/main@{#69008}
GitOrigin-RevId: f4e67517d598cc7a0b3d41c0a6751f840fbc6b06