siso: tweak siso_output explanation slightly
if the compile error incorrectly printed or cut off in the terminal,
then siso_output offers a way for developers to see what the full error
was. the wording we use to point people at this could be IMO tweaked
a bit to emphasize this.

Bug: b/381113092
Change-Id: I9b4dfe5b3824c7fdcb854d74427544a0225db0c0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/6063921
Reviewed-by: Philipp Wollermann <philwo@google.com>
Commit-Queue: Richard Wang <richardwa@google.com>
Cr-Commit-Position: refs/heads/main@{#70314}
GitOrigin-RevId: 3f1c451037615299f41a1d9f3a08363bc70432a1