siso: introduce "restat_content"
"restat=true" works nicely for cxx->solink->link.
e.g.
- when modified some source
- cxx step compiles the source and produces new *.o
- solink links *.o and produces *.so and *.so.TOC
  - it updates *.so (contains different bits of new *.o)
  - it may not update *.so.TOC if exported symbols are the same
- link steps depends on *.so.TOC, and use *.so on command line.
  - if exported symbols are not modified, no need to relink
    even if *.so is updated.
  - otherwise, need to relink

with remote-link, it won't work like that, because
- restat=true doesn't work for remote execution, since
  there is no mtime in output.
  we might be able to use node property, but it would not be good
  for exec cache
- for the case pruned by restat=true above (i.e. no modify on *.so.TOC),
  remote link step may not hit exec cache, because it uses different *.so
  as input.

To address this issue, introduce "restat_content" to consider
output is clean when content is not updated, like restat,
but not by mtime, but by content, and not considered as the same
if content is empty case, e.g. stamp file.
We may not be able to override "restat" because some step may produce
the same content but need to consider as modified when mtime differs.
e.g. LC_REEXPORT_DYLIB in apple link case.

Bug: b/381211851
Change-Id: I5c03eb72f811ebb3379a83f74e01a81375d28263
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/6170537
Commit-Queue: Fumitoshi Ukai <ukai@google.com>
Reviewed-by: Junji Watanabe <jwata@google.com>
Reviewed-by: Philipp Wollermann <philwo@google.com>
Auto-Submit: Fumitoshi Ukai <ukai@google.com>
Cr-Commit-Position: refs/heads/main@{#70973}
GitOrigin-RevId: cfc01604a2cbbfdd36999ae6d83152448a2c51a3