[siso] Slightly rewrite parseFlagsFully method.
The way this method is written is somewhat confusing. It contains an
unbounded loop although it is bounded in practice.

Regardless, intrange gets confused by the lifetime of `i`, which is
declared before and used after the for loop that uses it.

It's simple enough to use an `argsRemaining` variable to keep track of
the number of arguments that we didn't parse and this solves the
immediate problem that's confusing intrange (rewrite in later CLs).

Ideally, I would like to rewrite this method to be a little bit less
confusing and maybe put it in LUCI commons somewhere. The normal go CLI
convention of:

  thing subcommand -flag value -flag value -flag value positional-arguments

Is not very user-friendly and this causes lots of headaches in real life
for projects like `shivas`. It is frequently more intuitive to put flags
after the positional arguments. Siso ninja is ahead of the curve here.

Change-Id: Ifd214a57995347dd9a11a8186589320e85fd66b8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/6295195
Reviewed-by: Fumitoshi Ukai <ukai@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#71862}
GitOrigin-RevId: f12133fecc125fd5cd5c2a28c1539ae706b038e5