siso: re-run step if build inputs/outputs are added or removed.
Ninja only checks command hash and inputs/output's timestamp
to decide if the step need to re-run or not.
When new dependencies are added to the step and not changed
command line, ninja doesn't re-run the step.

n2 uses manifest to decide if the step needs to re-run or not.
This CL adds edgeHash as hash of inputs/outputs paths to detect
the dependencies change.

Bug: b/374177226, b/412265194
Change-Id: I3303a433127c43c03f02e73343ac13da461e046e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/6576947
Reviewed-by: Junji Watanabe <jwata@google.com>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Fumitoshi Ukai <ukai@google.com>
Auto-Submit: Fumitoshi Ukai <ukai@google.com>
Cr-Commit-Position: refs/heads/main@{#73798}
GitOrigin-RevId: 218dc31109b1f821c47053c46c0e066058b4601d