[errors] Refactor uses of errors.Annotate where `err` could be nil.
This is finding all locations where `errors.Annotate` is called with
an error that MIGHT be `nil`. The replacements are:
  * `errors.WrapIf` - the err specifically could be nil, and we just
    want to do the `if err != nil { <wrap> }` thing.
  * `errors.Fmt` - the err is NOT ever nil, and we want to
    unconditionally format it.

This is for an upcoming change to remove `errors.Annotate` entirely,
but `errors.Fmt` (the automatic replacement for `errors.Annotate`)
will unconditionally return a non-nil error, even if it's told to
wrap `nil`.

R=dhanyaganesh@chromium.org, gatong@chromium.org, gregorynisbet@google.com, richardwa@google.com

Change-Id: Ic42c2729d09fc56d7b25ada94740e762d919c111
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/6657931
Commit-Queue: Julio Hurtado <juahurta@google.com>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Auto-Submit: Robbie Iannucci <iannucci@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#74428}
GitOrigin-RevId: 3e176d0c1a3ce75cc10b3c13ce889801e692d8b4